+++ src/Plugin.Settings.Android/Settings.cs
@@ -22,12 +22,13 @@ namespace Plugin.Settings
         /// <typeparam name="T">Vaue of t (bool, int, float, long, string)</typeparam>
         /// <param name="key">Key for settings</param>
         /// <param name="defaultValue">default value if not set</param>
+        /// <param name="fileName">Name of file for settings to be stored and retrieved (iOS = SuiteName, Android = Name, Windows Store/RT8.1/UWP = Container name, WinPhone 8 SL = Doesn't Apply)</param>
         /// <returns>Value or default</returns>
-        public T GetValueOrDefault<T>(string key, T defaultValue = default(T))
+        public T GetValueOrDefault<T>(string key, T defaultValue = default(T), string fileName = null)
         {
             lock (locker)
             {
-                using (var sharedPreferences = PreferenceManager.GetDefaultSharedPreferences(Application.Context))
+                using (var sharedPreferences = GetSharedPrefence(fileName))
                 {
                     return GetValueOrDefaultCore(sharedPreferences, key, defaultValue);
                 }
@@ -182,8 +183,9 @@ namespace Plugin.Settings
         /// </summary>
         /// <param name="key">key to update</param>
         /// <param name="value">value to set</param>
+        /// <param name="fileName">Name of file for settings to be stored and retrieved (iOS = SuiteName, Android = Name, Windows Store/RT8.1/UWP = Container name, WinPhone 8 SL = Doesn't Apply)</param>
         /// <returns>True if added or update and you need to save</returns>
-        public bool AddOrUpdateValue<T>(string key, T value)
+        public bool AddOrUpdateValue<T>(string key, T value, string fileName = null)
         {
             Type typeOf = typeof(T);
             if (typeOf.IsGenericType && typeOf.GetGenericTypeDefinition() == typeof(Nullable<>))
@@ -191,14 +193,14 @@ namespace Plugin.Settings
                 typeOf = Nullable.GetUnderlyingType(typeOf);
             }
             var typeCode = Type.GetTypeCode(typeOf);
-            return AddOrUpdateValue(key, value, typeCode);
+            return AddOrUpdateValue(key, value, typeCode, fileName);
         }
 
-        private bool AddOrUpdateValue(string key, object value, TypeCode typeCode)
+        private bool AddOrUpdateValue(string key, object value, TypeCode typeCode, string fileName)
         {
             lock (locker)
             {
-                using (var sharedPreferences = PreferenceManager.GetDefaultSharedPreferences(Application.Context))
+                using (var sharedPreferences = GetSharedPrefence(fileName))
                 {
                     using (var sharedPreferencesEditor = sharedPreferences.Edit())
                     {
@@ -259,11 +261,12 @@ namespace Plugin.Settings
         /// Removes a desired key from the settings
         /// </summary>
         /// <param name="key">Key for setting</param>
-        public void Remove(string key)
+        /// <param name="fileName">Name of file for settings to be stored and retrieved (iOS = SuiteName, Android = Name, Windows Store/RT8.1/UWP = Container name, WinPhone 8 SL = Doesn't Apply)</param>
+        public void Remove(string key, string fileName = null)
         {
             lock (locker)
             {
-                using (var sharedPreferences = PreferenceManager.GetDefaultSharedPreferences(Application.Context))
+                using (var sharedPreferences = GetSharedPrefence(fileName))
                 {
                     using (var sharedPreferencesEditor = sharedPreferences.Edit())
                     {
@@ -277,11 +280,12 @@ namespace Plugin.Settings
         /// <summary>
         /// Clear all keys from settings
         /// </summary>
-        public void Clear()
+        /// <param name="fileName">Name of file for settings to be stored and retrieved (iOS = SuiteName, Android = Name, Windows Store/RT8.1/UWP = Container name, WinPhone 8 SL = Doesn't Apply)</param>
+        public void Clear(string fileName = null)
         {
             lock (locker)
             {
-                using (var sharedPreferences = PreferenceManager.GetDefaultSharedPreferences(Application.Context))
+                using (var sharedPreferences = GetSharedPrefence(fileName))
                 {
                     using (var sharedPreferencesEditor = sharedPreferences.Edit())
                     {
@@ -296,16 +300,22 @@ namespace Plugin.Settings
         /// Checks to see if the key has been added.
         /// </summary>
         /// <param name="key">Key to check</param>
+        /// <param name="fileName">Name of file for settings to be stored and retrieved (iOS = SuiteName, Android = Name, Windows Store/RT8.1/UWP = Container name, WinPhone 8 SL = Doesn't Apply)</param>
         /// <returns>True if contains key, else false</returns>
-        public bool Contains(string key)
+        public bool Contains(string key, string fileName = null)
         {
             lock (locker)
             {
-                using (var sharedPreferences = PreferenceManager.GetDefaultSharedPreferences(Application.Context))
+                using (var sharedPreferences = GetSharedPrefence(fileName))
                 {
                     return sharedPreferences.Contains(key);
                 }
             }
         }
+
+        ISharedPreferences GetSharedPrefence(string fileName) =>
+            string.IsNullOrWhiteSpace(fileName) ?
+            PreferenceManager.GetDefaultSharedPreferences(Application.Context) :
+            Application.Context.GetSharedPreferences(fileName, FileCreationMode.Private);
     }
 }